bizzLogic almost done

+++ ProjSpringboot/rooms_scanner/src/main/java/ies/p1/rooms_scanner/Entities/Rooms.java
@@ -9,16 +9,16 @@ import javax.persistence.Table;
 @Entity
 @Table(name = "rooms")
 public class Rooms {
-    @Id     // ->  !!!!!!!!!!! TODO: ID automatico ou dept+floor+numSala?
+    @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
+    private int id;
     private String department;
     private int busySeats;
     private int maxSeats;
     private int floor;
     // room(id, departamento, andar, numLugaresOcupados, maxSeats)
     public Rooms() {}
-    public long getId() {
+    public int getId() {
         return id;
     }
 
@@ -30,7 +30,7 @@ public class Rooms {
         this.floor = floor;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
